+++ include/common.h
@@ -35,7 +35,7 @@ const int kMaxSeqName;
 
 void de_verbose(char const *fmt, ...);
 void de_debug(char const *fmt, ...);
-void* de_malloc(int32_t n);
+void* de_malloc(size_t n);
 int32_t de_getline(char **s, int32_t *n, FILE *f);
 FILE* de_fopen(const char *s, char const *mode);
 char* de_strdup(const char *s);
